Add bvec reorientation option
adds a --bvec option, allowing the user to specify a bvec file accompanying an image they are reorienting. This option will reorient the bvec file into the desired orientation and write to a new file with the same  base name as the output image, but with the .bvec extension.

vtkAnatomicalOrientation is a new class included in the depends for Utilities.